move Instr to translator utils
